Update README and rename enabled setting to format
